fix: use uv path for managed python installs
I had mistakenly believed that `process.resourcesPath` was a persistent directory and downloaded our portable python to this directory. This is actually an ephemeral temp directory.

As a result, the python installed when installing Invoke via launcher was deleted when the launcher closed.

Due to a fluke of dev environment setup, this problem wasn't caught.

This commit walks back the attempt to store our own portable python separately from uv's default install locations. It should resolve issues where the launcher fails to find or remember the user's installation.