feat: support manual version entry
Another option is added to the version picker for manual version entry. When selected, a free-text entry field is presented.

When manual version entry is used, the user is warned that manual version installation can break things and to back up their data to be safe.

The version is checked against PEP440 to ensure it is a valid version string, but we do not query PyPI to confirm it is a valid version for Invoke. If it is not a valid version, the install will simply fail.

A future enhancement could add validation against PyPI.