experiment: migrate from electron-forge to electron-vite for build + conveyor for packaging
Not fully working.

Notes:
- Conveyor can build all platforms from any platform. This is
wonderfully convenient, but we need to build node-pty on each platform,
so we cannot take advantage of the cross-platform build functionality.
  - https://github.com/microsoft/node-pty/
- Conveyor wants to inject something into native binaries on macOS to
support signing and/or auto-updater functionality. It seems to want to
do this to the node-pty binary. Cannot figure out how to opt out of this
behaviour. It results in an error like this:
  - ❌ Task failed: Mac App for Intel: Signing Mac binary pty.o.tmp
failed. Please ensure the binary was linked using the `-headerpad 0xFF`
flag to Apple's ld.`
  - See https://conveyor.hydraulic.dev/19.0/configs/native-apps/#macos
- I tried to disable signing but it seems to be impossible. Conveyor
seems to always sign, using a self-signed cert when "proper" signing
creds are not provided.
- Conveyor _does_ support signing w/ DigiCert's cloud signing service
that we use. It does this while packaging. I have not attempted to set
it up. There is a note on the docs about excessive usage of signing
quota for cloud signing services:
  -
https://conveyor.hydraulic.dev/19.0/faq/signing-and-certificates/#11-i-did-a-build-and-my-windows-cloud-signing-service-charged-me-for-hundreds-of-signatures
- Conveyor suggests using the Windows Store as a kind of hack to get the
app signed for cheap. It costs $19 for the sub and MS signs it for free.
Then I suppose you still get the ability to distribute manually.
- Conveyor is closed source, but free for OSS. It seems to be not
particularly popular among OSS, but I saw several posts about people who
were very pleased with its ease of use. Who knows if it is destined to
be well-supported long-term.